Move changes to merchandise
