+++ website/merchandise/tasks.py
@@ -5,4 +5,5 @@ from . import services
 
 @shared_task
 def renew_merchandise_sale_shift():
-    services.renew_merchandise_sale_shift()
+    services.lock_merchandise_sale_shift()
+    services.create_daily_merchandise_sale_shift()
