+++ website/sales/apps.py
@@ -5,7 +5,6 @@ class SalesConfig(AppConfig):
     name = "sales"
 
     def ready(self):
-        from . import signals  # noqa: F401
         from .payables import register
 
         register()
